Admin formatting andone class per file
