+++ code/Admin/ShopAdmin.php
@@ -2,9 +2,9 @@
 
 namespace SwipeStripe\Core\Admin;
 
-use SwipeStripe\Core\code\Order\Order;
-use SwipeStripe\Core\code\Customer\Customer;
-use SwipeStripe\Core\code\Admin\ShopConfig;
+use SwipeStripe\Core\Order\Order;
+use SwipeStripe\Core\Customer\Customer;
+use SwipeStripe\Core\Admin\ShopConfig;
 use SilverStripe\View\Requirements;
 use SilverStripe\ORM\ArrayList;
 use SilverStripe\View\ArrayData;
@@ -17,7 +17,6 @@ use SilverStripe\Forms\GridField\GridFieldDetailForm;
 use SilverStripe\Forms\FieldList;
 use SilverStripe\Forms\Form;
 use SilverStripe\Control\Controller;
-use SwipeStripe\Core\code\Admin\ShopAdmin;
 use SilverStripe\Core\ClassInfo;
 use SilverStripe\Core\Config\Config;
 use SilverStripe\Admin\ModelAdmin;
@@ -33,7 +32,7 @@ use SilverStripe\Security\Permission;
 use SilverStripe\Forms\NumericField;
 use SilverStripe\Forms\LiteralField;
 use SilverStripe\Forms\RequiredFields;
-use SwipeStripe\Core\code\Product\Attribute;
+use SwipeStripe\Core\Product\Attribute;
 use SilverStripe\Core\Extension;
 
 /**
@@ -54,43 +53,42 @@ class ShopAdmin extends ModelAdmin
 
     public $showImportForm = false;
 
-    private static $managed_models = array(
+    private static $managed_models = [
         //'Product',
         Order::class,
         Customer::class,
         ShopConfig::class
-    );
+    ];
 
-    private static $url_handlers = array(
+    private static $url_handlers = [
         '$ModelClass/$Action' => 'handleAction',
         '$ModelClass/$Action/$ID' => 'handleAction',
-    );
+    ];
 
-    public static $hidden_sections = array();
-    
-    private static $allowed_actions = array(
+    public static $hidden_sections = [];
+
+    private static $allowed_actions = [
         'EditForm',
         'SettingsContent',
         'SettingsForm'
-    );
+    ];
 
     public function init()
     {
-
         // set reading lang
         // if(Object::has_extension('SiteTree', 'Translatable') && !$this->request->isAjax()) {
-        // 	Translatable::choose_site_locale(array_keys(Translatable::get_existing_content_languages('SiteTree')));
+        // Translatable::choose_site_locale(array_keys(Translatable::get_existing_content_languages('SiteTree')));
         // }
-        
+
         parent::init();
 
         Requirements::css(CMS_DIR . '/css/screen.css');
         Requirements::css('swipestripe/css/ShopAdmin.css');
-        
+
         Requirements::combine_files(
             'cmsmain.js',
             array_merge(
-                array(
+                [
                     CMS_DIR . '/javascript/CMSMain.js',
                     CMS_DIR . '/javascript/CMSMain.EditForm.js',
                     CMS_DIR . '/javascript/CMSMain.AddForm.js',
@@ -98,7 +96,7 @@ class ShopAdmin extends ModelAdmin
                     CMS_DIR . '/javascript/CMSMain.Tree.js',
                     CMS_DIR . '/javascript/SilverStripeNavigator.js',
                     CMS_DIR . '/javascript/SiteTreeURLSegmentField.js'
-                ),
+                ],
                 Requirements::add_i18n_javascript(CMS_DIR . '/javascript/lang', true, true)
             )
         );
@@ -118,7 +116,7 @@ class ShopAdmin extends ModelAdmin
     {
         $models = $this->stat('managed_models');
         if (is_string($models)) {
-            $models = array($models);
+            $models = [$models];
         }
         if (!count($models)) {
             user_error(
@@ -132,7 +130,7 @@ class ShopAdmin extends ModelAdmin
         // Normalize models to have their model class in array key
         foreach ($models as $k => $v) {
             if (is_numeric($k)) {
-                $models[$v] = array('title' => singleton($v)->i18n_plural_name());
+                $models[$v] = ['title' => singleton($v)->i18n_plural_name()];
                 unset($models[$k]);
             }
         }
@@ -147,18 +145,18 @@ class ShopAdmin extends ModelAdmin
      */
     protected function getManagedModelTabs()
     {
-        $forms  = new ArrayList();
+        $forms = new ArrayList();
 
         $models = $this->getManagedModels();
         foreach ($models as $class => $options) {
-            $forms->push(new ArrayData(array(
-                'Title'     => $options['title'],
+            $forms->push(new ArrayData([
+                'Title' => $options['title'],
                 'ClassName' => $class,
                 'Link' => $this->Link($this->sanitiseClassName($class)),
                 'LinkOrCurrent' => ($class == $this->modelClass) ? 'current' : 'link'
-            )));
+            ]));
         }
-        
+
         return $forms;
     }
 
@@ -183,12 +181,11 @@ class ShopAdmin extends ModelAdmin
 
     public function getEditForm($id = null, $fields = null)
     {
-
         //If editing the shop settings get the first back and edit that basically...
         if ($this->modelClass == ShopConfig::class) {
             return $this->renderWith('ShopAdmin_ConfigEditForm');
         }
-        
+
         $list = $this->getList();
 
         $buttonAfter = new GridFieldButtonRow('after');
@@ -228,7 +225,7 @@ class ShopAdmin extends ModelAdmin
         $form->setAttribute('data-pjax-fragment', 'CurrentForm');
 
         $this->extend('updateEditForm', $form);
-        
+
         return $form;
     }
 
@@ -247,7 +244,7 @@ class ShopAdmin extends ModelAdmin
         $snippets = new ArrayList();
         $subClasses = ClassInfo::subclassesFor(ShopAdmin::class);
 
-        $classes = array();
+        $classes = [];
         foreach ($subClasses as $className) {
             $classes[$className] = Config::inst()->get($className, 'url_priority');
         }
@@ -258,9 +255,9 @@ class ShopAdmin extends ModelAdmin
             $snippet = $obj->getSnippet();
 
             if ($snippet && !in_array($className, self::$hidden_sections)) {
-                $snippets->push(new ArrayData(array(
+                $snippets->push(new ArrayData([
                     'Content' => $snippet
-                )));
+                ]));
             }
         }
         return $snippets;
@@ -269,594 +266,5 @@ class ShopAdmin extends ModelAdmin
     public function getSnippet()
     {
         return false;
-    }
-}
-
-/**
- * Shop admin area for managing email settings
- *
- * @author Frank Mullenger <frankmullenger@gmail.com>
- * @copyright Copyright (c) 2011, Frank Mullenger
- * @package swipestripe
- * @subpackage admin
- */
-class ShopAdmin_EmailAdmin extends ShopAdmin
-{
-    private static $tree_class = ShopConfig::class;
-    
-    private static $allowed_actions = array(
-        'EmailSettings',
-        'EmailSettingsForm',
-        'saveEmailSettings'
-    );
-
-    private static $url_rule = 'ShopConfig/EmailSettings';
-    private static $url_priority = 60;
-    private static $menu_title = 'Shop Emails';
-
-    private static $url_handlers = array(
-        'ShopConfig/EmailSettings/EmailSettingsForm' => 'EmailSettingsForm',
-        'ShopConfig/EmailSettings' => 'EmailSettings'
-    );
-
-    public function init()
-    {
-        parent::init();
-        if (!in_array(get_class($this), self::$hidden_sections)) {
-            $this->modelClass = ShopConfig::class;
-        }
-    }
-
-    public function Breadcrumbs($unlinked = false)
-    {
-        $request = $this->getRequest();
-        $items = parent::Breadcrumbs($unlinked);
-
-        if ($items->count() > 1) {
-            $items->remove($items->pop());
-        }
-
-        $items->push(new ArrayData(array(
-            'Title' => 'Email Settings',
-            'Link' => $this->Link(Controller::join_links($this->sanitiseClassName($this->modelClass), 'EmailSettings'))
-        )));
-
-        return $items;
-    }
-
-    public function SettingsForm($request = null)
-    {
-        return $this->EmailSettingsForm();
-    }
-
-    public function EmailSettings($request)
-    {
-        if ($request->isAjax()) {
-            $controller = $this;
-            $responseNegotiator = new PjaxResponseNegotiator(
-                array(
-                    'CurrentForm' => function () use (&$controller) {
-                        return $controller->EmailSettingsForm()->forTemplate();
-                    },
-                    'Content' => function () use (&$controller) {
-                        return $controller->renderWith('ShopAdminSettings_Content');
-                    },
-                    'Breadcrumbs' => function () use (&$controller) {
-                        return $controller->renderWith('CMSBreadcrumbs');
-                    },
-                    'default' => function () use (&$controller) {
-                        return $controller->renderWith($controller->getViewer('show'));
-                    }
-                ),
-                $this->response
-            );
-            return $responseNegotiator->respond($this->getRequest());
-        }
-
-        return $this->renderWith('ShopAdminSettings');
-    }
-
-    public function EmailSettingsForm()
-    {
-        $shopConfig = ShopConfig::get()->First();
-
-        $fields = new FieldList(
-            $rootTab = new TabSet(
-                "Root",
-                $tabMain = new Tab(
-                    'Receipt',
-                    new HiddenField('ShopConfigSection', null, 'EmailSettings'),
-                    new TextField('ReceiptFrom', _t('ShopConfig.FROM', 'From')),
-                    TextField::create('ReceiptTo', _t('ShopConfig.TO', 'To'))
-                        ->setValue(_t('ShopConfig.RECEIPT_TO', 'Sent to customer'))
-                        ->performReadonlyTransformation(),
-                    new TextField('ReceiptSubject', _t('ShopConfig.SUBJECT_LINE', 'Subject line')),
-                    TextareaField::create('ReceiptBody', _t('ShopConfig.MESSAGE', 'Message'))
-                        ->setRightTitle(_t('ShopConfig.MESSAGE_DETAILS', 'Order details are included in the email below this message')),
-                    new TextareaField('EmailSignature', _t('ShopConfig.SIGNATURE', 'Signature'))
-                ),
-                new Tab(
-                    'Notification',
-                    TextField::create('NotificationFrom', _t('ShopConfig.FROM', 'From'))
-                        ->setValue(_t('ShopConfig.NOTIFICATION_FROM', 'Customer email address'))
-                        ->performReadonlyTransformation(),
-                    new TextField('NotificationTo', _t('ShopConfig.TO', 'To')),
-                    new TextField('NotificationSubject', _t('ShopConfig.SUBJECT_LINE', 'Subject line')),
-                    TextareaField::create('NotificationBody', _t('ShopConfig.MESSAGE', 'Message'))
-                        ->setRightTitle(_t('ShopConfig.MESSAGE_DETAILS', 'Order details are included in the email below this message'))
-                )
-            )
-        );
-
-        $actions = new FieldList();
-        $actions->push(FormAction::create('saveEmailSettings', _t('GridFieldDetailForm.Save', 'Save'))
-            ->setUseButtonTag(true)
-            ->addExtraClass('ss-ui-action-constructive')
-            ->setAttribute('data-icon', 'add'));
-
-        $form = new Form(
-            $this,
-            'EditForm',
-            $fields,
-            $actions
-        );
-
-        $form->setTemplate('ShopAdminSettings_EditForm');
-        $form->setAttribute('data-pjax-fragment', 'CurrentForm');
-        $form->addExtraClass('cms-content cms-edit-form center ss-tabset');
-        if ($form->Fields()->hasTabset()) {
-            $form->Fields()->findOrMakeTab('Root')->setTemplate('CMSTabSet');
-        }
-        $form->setFormAction(Controller::join_links($this->Link($this->sanitiseClassName($this->modelClass)), 'EmailSettings/EmailSettingsForm'));
-
-        $form->loadDataFrom($shopConfig);
-
-        return $form;
-    }
-
-    public function saveEmailSettings($data, $form)
-    {
-
-        //Hack for LeftAndMain::getRecord()
-        // self::$tree_class = 'ShopConfig';
-
-        $config = ShopConfig::get()->First();
-        $form->saveInto($config);
-        $config->write();
-        $form->sessionMessage('Saved Email Settings', 'good');
-
-        $controller = $this;
-        $responseNegotiator = new PjaxResponseNegotiator(
-            array(
-                'CurrentForm' => function () use (&$controller) {
-                    //return $controller->renderWith('ShopAdminSettings_Content');
-                    return $controller->EmailSettingsForm()->forTemplate();
-                },
-                'Content' => function () use (&$controller) {
-                    //return $controller->renderWith($controller->getTemplatesWithSuffix('_Content'));
-                },
-                'Breadcrumbs' => function () use (&$controller) {
-                    return $controller->renderWith('CMSBreadcrumbs');
-                },
-                'default' => function () use (&$controller) {
-                    return $controller->renderWith($controller->getViewer('show'));
-                }
-            ),
-            $this->response
-        );
-        return $responseNegotiator->respond($this->getRequest());
-    }
-
-    public function getSnippet()
-    {
-        if (!$member = Member::currentUser()) {
-            return false;
-        }
-        if (!Permission::check('CMS_ACCESS_' . get_class($this), 'any', $member)) {
-            return false;
-        }
-
-        return $this->customise(array(
-            'Title' => 'Email Settings',
-            'Help' => 'Order notification and receipt details and recipeients.',
-            'Link' => Controller::join_links($this->Link(ShopConfig::class), 'EmailSettings'),
-            'LinkTitle' => 'Edit Email Settings'
-        ))->renderWith('ShopAdmin_Snippet');
-    }
-}
-
-/**
- * Shop admin area for managing base currency
- *
- * @author Frank Mullenger <frankmullenger@gmail.com>
- * @copyright Copyright (c) 2011, Frank Mullenger
- * @package swipestripe
- * @subpackage admin
- */
-class ShopAdmin_BaseCurrency extends ShopAdmin
-{
-    private static $tree_class = ShopConfig::class;
-    
-    private static $allowed_actions = array(
-        'BaseCurrencySettings',
-        'BaseCurrencySettingsForm',
-        'saveBaseCurrencySettings'
-    );
-
-    private static $url_rule = 'ShopConfig/BaseCurrency';
-    private static $url_priority = 65;
-    private static $menu_title = 'Shop Base Currency';
-
-    private static $url_handlers = array(
-        'ShopConfig/BaseCurrency/BaseCurrencySettingsForm' => 'BaseCurrencySettingsForm',
-        'ShopConfig/BaseCurrency' => 'BaseCurrencySettings'
-    );
-
-    public function init()
-    {
-        parent::init();
-        if (!in_array(get_class($this), self::$hidden_sections)) {
-            $this->modelClass = ShopConfig::class;
-        }
-    }
-
-    public function Breadcrumbs($unlinked = false)
-    {
-        $request = $this->getRequest();
-        $items = parent::Breadcrumbs($unlinked);
-
-        if ($items->count() > 1) {
-            $items->remove($items->pop());
-        }
-
-        $items->push(new ArrayData(array(
-            'Title' => 'Base Currency',
-            'Link' => $this->Link(Controller::join_links($this->sanitiseClassName($this->modelClass), 'BaseCurrency'))
-        )));
-
-        return $items;
-    }
-
-    public function SettingsForm($request = null)
-    {
-        return $this->BaseCurrencySettingsForm();
-    }
-
-    public function BaseCurrencySettings($request)
-    {
-        if ($request->isAjax()) {
-            $controller = $this;
-            $responseNegotiator = new PjaxResponseNegotiator(
-                array(
-                    'CurrentForm' => function () use (&$controller) {
-                        return $controller->BaseCurrencySettingsForm()->forTemplate();
-                    },
-                    'Content' => function () use (&$controller) {
-                        return $controller->renderWith('ShopAdminSettings_Content');
-                    },
-                    'Breadcrumbs' => function () use (&$controller) {
-                        return $controller->renderWith('CMSBreadcrumbs');
-                    },
-                    'default' => function () use (&$controller) {
-                        return $controller->renderWith($controller->getViewer('show'));
-                    }
-                ),
-                $this->response
-            );
-            return $responseNegotiator->respond($this->getRequest());
-        }
-
-        return $this->renderWith('ShopAdminSettings');
-    }
-
-    public function BaseCurrencySettingsForm()
-    {
-        $shopConfig = ShopConfig::get()->First();
-
-        $fields = new FieldList(
-            $rootTab = new TabSet(
-                "Root",
-                $tabMain = new Tab(
-                    'BaseCurrency',
-                    TextField::create('BaseCurrency', _t('ShopConfig.BASE_CURRENCY', 'Base Currency'))
-                        ->setRightTitle('3 letter code for base currency - <a href="http://en.wikipedia.org/wiki/ISO_4217#Active_codes" target="_blank">available codes</a>'),
-                    TextField::create('BaseCurrencySymbol', _t('ShopConfig.BASE_CURRENCY_SYMBOL', 'Base Currency Symbol'))
-                        ->setRightTitle('Symbol to be used for the base currency e.g: $'),
-                    NumericField::create('BaseCurrencyPrecision', _t('ShopConfig.BASE_CURRENCY_PRECISION', 'Base Currency Precision'))
-                        ->setRightTitle('Most currencies use two digits after the decimal place. If using digital currencies like Bitcoin, precision should be at least eight digits after the decimal.')
-                )
-            )
-        );
-
-        if ($shopConfig->BaseCurrency) {
-            $fields->addFieldToTab('Root.BaseCurrency', new LiteralField('BaseCurrencyNotice', '
-				<p class="message warning">Base currency has already been set, do not change unless you know what you are doing.</p>
-			'), 'BaseCurrency');
-        }
-
-        $actions = new FieldList();
-        $actions->push(FormAction::create('saveBaseCurrencySettings', _t('GridFieldDetailForm.Save', 'Save'))
-            ->setUseButtonTag(true)
-            ->addExtraClass('ss-ui-action-constructive')
-            ->setAttribute('data-icon', 'add'));
-
-        $validator = new RequiredFields('BaseCurrency');
-
-        $form = new Form(
-            $this,
-            'EditForm',
-            $fields,
-            $actions,
-            $validator
-        );
-
-        $form->setTemplate('ShopAdminSettings_EditForm');
-        $form->setAttribute('data-pjax-fragment', 'CurrentForm');
-        $form->addExtraClass('cms-content cms-edit-form center ss-tabset');
-        if ($form->Fields()->hasTabset()) {
-            $form->Fields()->findOrMakeTab('Root')->setTemplate('CMSTabSet');
-        }
-        $form->setFormAction(Controller::join_links($this->Link($this->sanitiseClassName($this->modelClass)), 'BaseCurrency/BaseCurrencySettingsForm'));
-
-        $form->loadDataFrom($shopConfig);
-
-        return $form;
-    }
-
-    public function saveBaseCurrencySettings($data, $form)
-    {
-
-        //Hack for LeftAndMain::getRecord()
-        self::$tree_class = ShopConfig::class;
-
-        $config = ShopConfig::get()->First();
-        $form->saveInto($config);
-        $config->write();
-        $form->sessionMessage('Saved BaseCurrency Key', 'good');
-
-        $controller = $this;
-        $responseNegotiator = new PjaxResponseNegotiator(
-            array(
-                'CurrentForm' => function () use (&$controller) {
-                    //return $controller->renderWith('ShopAdminSettings_Content');
-                    return $controller->BaseCurrencySettingsForm()->forTemplate();
-                },
-                'Content' => function () use (&$controller) {
-                    //return $controller->renderWith($controller->getTemplatesWithSuffix('_Content'));
-                },
-                'Breadcrumbs' => function () use (&$controller) {
-                    return $controller->renderWith('CMSBreadcrumbs');
-                },
-                'default' => function () use (&$controller) {
-                    return $controller->renderWith($controller->getViewer('show'));
-                }
-            ),
-            $this->response
-        );
-        return $responseNegotiator->respond($this->getRequest());
-    }
-
-    public function getSnippet()
-    {
-        if (!$member = Member::currentUser()) {
-            return false;
-        }
-        if (!Permission::check('CMS_ACCESS_' . get_class($this), 'any', $member)) {
-            return false;
-        }
-
-        return $this->customise(array(
-            'Title' => 'Base Currency',
-            'Help' => 'Set base currency.',
-            'Link' => Controller::join_links($this->Link(ShopConfig::class), 'BaseCurrency'),
-            'LinkTitle' => 'Edit base currency'
-        ))->renderWith('ShopAdmin_Snippet');
-    }
-}
-
-/**
- * Shop admin area for managing product attributes.
- *
- * @author Frank Mullenger <frankmullenger@gmail.com>
- * @copyright Copyright (c) 2011, Frank Mullenger
- * @package swipestripe
- * @subpackage admin
- */
-class ShopAdmin_Attribute extends ShopAdmin
-{
-    private static $tree_class = ShopConfig::class;
-    
-    private static $allowed_actions = array(
-        'AttributeSettings',
-        'AttributeSettingsForm',
-        'saveAttributeSettings'
-    );
-
-    private static $url_rule = 'ShopConfig/Attribute';
-    private static $url_priority = 75;
-    private static $menu_title = 'Shop Product Attributes';
-
-    private static $url_handlers = array(
-        'ShopConfig/Attribute/AttributeSettingsForm' => 'AttributeSettingsForm',
-        'ShopConfig/Attribute' => 'AttributeSettings'
-    );
-
-    public function init()
-    {
-        parent::init();
-        if (!in_array(get_class($this), self::$hidden_sections)) {
-            $this->modelClass = ShopConfig::class;
-        }
-    }
-
-    public function Breadcrumbs($unlinked = false)
-    {
-        $request = $this->getRequest();
-        $items = parent::Breadcrumbs($unlinked);
-
-        if ($items->count() > 1) {
-            $items->remove($items->pop());
-        }
-
-        $items->push(new ArrayData(array(
-            'Title' => 'Attribute Settings',
-            'Link' => $this->Link(Controller::join_links($this->sanitiseClassName($this->modelClass), Attribute::class))
-        )));
-
-        return $items;
-    }
-
-    public function SettingsForm($request = null)
-    {
-        return $this->AttributeSettingsForm();
-    }
-
-    public function AttributeSettings($request)
-    {
-        if ($request->isAjax()) {
-            $controller = $this;
-            $responseNegotiator = new PjaxResponseNegotiator(
-                array(
-                    'CurrentForm' => function () use (&$controller) {
-                        return $controller->AttributeSettingsForm()->forTemplate();
-                    },
-                    'Content' => function () use (&$controller) {
-                        return $controller->renderWith('ShopAdminSettings_Content');
-                    },
-                    'Breadcrumbs' => function () use (&$controller) {
-                        return $controller->renderWith('CMSBreadcrumbs');
-                    },
-                    'default' => function () use (&$controller) {
-                        return $controller->renderWith($controller->getViewer('show'));
-                    }
-                ),
-                $this->response
-            );
-            return $responseNegotiator->respond($this->getRequest());
-        }
-
-        return $this->renderWith('ShopAdminSettings');
-    }
-
-    public function AttributeSettingsForm()
-    {
-        $shopConfig = ShopConfig::get()->First();
-
-        $fields = new FieldList(
-            $rootTab = new TabSet(
-                'Root',
-                $tabMain = new Tab(
-                    Attribute::class,
-                    GridField::create(
-                        'Attributes',
-                        'Attributes',
-                        $shopConfig->Attributes(),
-                        GridFieldConfig_HasManyRelationEditor::create()
-                    )
-                )
-            )
-        );
-
-        $actions = new FieldList();
-        $actions->push(FormAction::create('saveAttributeSettings', _t('GridFieldDetailForm.Save', 'Save'))
-            ->setUseButtonTag(true)
-            ->addExtraClass('ss-ui-action-constructive')
-            ->setAttribute('data-icon', 'add'));
-
-        $form = new Form(
-            $this,
-            'EditForm',
-            $fields,
-            $actions
-        );
-
-        $form->setTemplate('ShopAdminSettings_EditForm');
-        $form->setAttribute('data-pjax-fragment', 'CurrentForm');
-        $form->addExtraClass('cms-content cms-edit-form center ss-tabset');
-        if ($form->Fields()->hasTabset()) {
-            $form->Fields()->findOrMakeTab('Root')->setTemplate('CMSTabSet');
-        }
-        $form->setFormAction(Controller::join_links($this->Link($this->sanitiseClassName($this->modelClass)), 'Attribute/AttributeSettingsForm'));
-
-        $form->loadDataFrom($shopConfig);
-
-        return $form;
-    }
-
-    public function saveAttributeSettings($data, $form)
-    {
-
-        //Hack for LeftAndMain::getRecord()
-        self::$tree_class = ShopConfig::class;
-
-        $config = ShopConfig::get()->First();
-        $form->saveInto($config);
-        $config->write();
-        $form->sessionMessage('Saved Attribute Settings', 'good');
-
-        $controller = $this;
-        $responseNegotiator = new PjaxResponseNegotiator(
-            array(
-                'CurrentForm' => function () use (&$controller) {
-                    //return $controller->renderWith('ShopAdminSettings_Content');
-                    return $controller->AttributeSettingsForm()->forTemplate();
-                },
-                'Content' => function () use (&$controller) {
-                    //return $controller->renderWith($controller->getTemplatesWithSuffix('_Content'));
-                },
-                'Breadcrumbs' => function () use (&$controller) {
-                    return $controller->renderWith('CMSBreadcrumbs');
-                },
-                'default' => function () use (&$controller) {
-                    return $controller->renderWith($controller->getViewer('show'));
-                }
-            ),
-            $this->response
-        );
-        return $responseNegotiator->respond($this->getRequest());
-    }
-
-    public function getSnippet()
-    {
-        if (!$member = Member::currentUser()) {
-            return false;
-        }
-        if (!Permission::check('CMS_ACCESS_' . get_class($this), 'any', $member)) {
-            return false;
-        }
-
-        return $this->customise(array(
-            'Title' => 'Attribute Management',
-            'Help' => 'Create default attributes',
-            'Link' => Controller::join_links($this->Link(ShopConfig::class), Attribute::class),
-            'LinkTitle' => 'Edit default attributes'
-        ))->renderWith('ShopAdmin_Snippet');
-    }
-}
-
-/**
- * Extension for admin area to apply shop admin CSS etc.
- *
- * @author Frank Mullenger <frankmullenger@gmail.com>
- * @copyright Copyright (c) 2011, Frank Mullenger
- * @package swipestripe
- * @subpackage admin
- */
-class ShopAdmin_LeftAndMainExtension extends Extension
-{
-    public function onAfterInit()
-    {
-        Requirements::css('swipestripe/css/ShopAdmin.css');
-    }
-
-    public function alternateMenuDisplayCheck($className)
-    {
-        if (class_exists($className)) {
-            $obj = new $className();
-            if (is_subclass_of($obj, ShopAdmin::class)) {
-                return false;
-            }
-        }
-        return true;
     }
 }
