+++ code/Admin/ShopSearchContext.php
@@ -16,7 +16,7 @@ class ShopSearchContext_Order extends SearchContext
 {
     public function getResults($searchParams, $sort = false, $limit = false)
     {
-        $searchParams = array_filter((array)$searchParams, array($this,'clearEmptySearchFields'));
+        $searchParams = array_filter((array)$searchParams, [$this, 'clearEmptySearchFields']);
 
         // getQuery actually returns a DataList
         $query = $this->getQuery($searchParams, $sort, $limit);
@@ -24,7 +24,7 @@ class ShopSearchContext_Order extends SearchContext
         //Only orders which have been processed are displayed
         $query = $query->leftJoin(
             $table = 'Payment',
-            $onPredicate = "\"Payment\".\"OrderID\" = \"Order\".\"ID\"",
+            $onPredicate = '"Payment"."OrderID" = "Order"."ID"',
             $tableAlias = 'Payment'
         );
         $query = $query->where('"Payment"."ID" IS NOT NULL');
